+++ src/app/home/page.tsx
@@ -2,11 +2,11 @@
 
 import Header from '@/feature/Header/Header';
 import ProfileCard from '@/feature/Profile/ProfileCard';
-import GoLoginButton from '@/feature/GoLogin/gologinbutton';
+import GoLoginButton from '@/feature/GoLogin/GoLoginButton';
 import StartButtons from '@/feature/StartButtons/StartButtons';
-import { useUserSummary } from '@/feature/Hooks/useUserSummary';
+import { useUserSummary } from '@/feature/hooks/useUserSummary';
 import LogoutButton from '@/feature/Logout/LogoutButton';
-import CrystalText from '@/feature/CrystalGrowText/crystalgrowtext';
+import CrystalText from '@/feature/CrystalGrowText/CrystalGrowText';
 
 export default function HomePage() {
   const { data, loading, error } = useUserSummary();
