+++ src/app/page.tsx
@@ -1,5 +1,5 @@
-import GoLoginButton from "@/feature/GoLogin/gologinbutton";
-import GoRegisterButton from "@/feature/GoRegister/goregisterbutton";
+import GoLoginButton from "@/feature/GoLogin/GoLoginButton";
+import GoRegisterButton from "@/feature/GoRegister/GoRegisterButton";
 import Image from "next/image"
 
 function Home() {
